Simplify GetSizeAsync, clean up logging and imports
Removed unnecessary casts in GetSizeAsync methods across storage classes and tests. Improved logging clarity in jobs and refactored code formatting. Removed redundant using directives to enhance code readability and maintainability.

+++ src/Cotton.Benchmark/Benchmarks/PipelineBenchmark.cs
@@ -126,7 +126,7 @@ namespace Cotton.Benchmark.Benchmarks
 
             public Task<long> GetSizeAsync(string uid)
             {
-                return Task.FromResult(_storage.TryGetValue(uid, out var data) ? (long)data.Length : 0L);
+                return Task.FromResult(_storage.TryGetValue(uid, out var data) ? data.Length : 0L);
             }
 
             public Task<Stream> ReadAsync(string uid)
