+++ src/Cotton.Storage.Tests/Pipelines/FileStoragePipelineTests.cs
@@ -33,7 +33,7 @@ namespace Cotton.Storage.Tests.Pipelines
 
             public Task<long> GetSizeAsync(string uid)
             {
-                return Task.FromResult(_storage.TryGetValue(uid, out var data) ? (long)data.Length : 0L);
+                return Task.FromResult(_storage.TryGetValue(uid, out var data) ? data.Length : 0L);
             }
 
             public Task<Stream> ReadAsync(string uid)
