+++ src/Cotton.Storage.Tests/Streams/ConcatenatedReadStreamTests.cs
@@ -33,7 +33,7 @@ namespace Cotton.Storage.Tests.Streams
 
             public Task<long> GetSizeAsync(string uid)
             {
-                return Task.FromResult(_data.TryGetValue(uid, out var data) ? (long)data.Length : 0L);
+                return Task.FromResult(_data.TryGetValue(uid, out var data) ? data.Length : 0L);
             }
 
             public Task<Stream> ReadAsync(string uid, PipelineContext? context = null)
